requirements.txt
customtkinter>=5.2.0
python-docx>=1.1.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
